Fix add_edge and change console printing
Changed the add_edge to check the vertex as well as the value to be connected with the vertex. Also
changed the printing of the visted to print each time a node is added to the visited and made the
visited be a set again.
Also got rid of comments and print statements in graph_demo.